refactor(/domain): 도메인 내부 패키지 정리
