+++ src/main/java/blackjack/domain/participant/Player.java
@@ -1,5 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
+import blackjack.domain.card.Deck;
+import blackjack.domain.game.GameResult;
+import blackjack.domain.game.Score;
 import blackjack.domain.card.Card;
 import java.util.List;
